moved victorias card+button from /home to /resume. starting work on upload docx functionality
