+++ src/app/home/page.tsx
@@ -5,7 +5,6 @@ import { useAuth } from "@/context/authContext";
 import { useRouter } from "next/navigation";
 import { useState, useEffect } from "react";
 import { Card, CardHeader, CardTitle, CardContent, CardDescription, CardFooter } from "@/components/ui/card";
-import SubmissionFeedback from "@/components/SubmissionFeedback";
 import { UserNameAddUpdate } from '@/components/userNameAddUpdate';
 
 
@@ -23,9 +22,6 @@ export default function HomePage() {
     // Debugging:
     const [message, setMessage] = useState('');
 
-    const [submissionStatus, setSubmissionStatus] = useState<null | { type: "success" | "error"; message: string }>(null);
-
-
 
     useEffect(() => {
         if (!loading && !user) {
@@ -36,67 +32,11 @@ export default function HomePage() {
     if (loading) {
         return <p>Loading...</p>; // Show a loading state while checking auth
     }
-
-    const handleSubmit = () => {
-        setSubmissionStatus(null); // clear current state first
-      
-        setTimeout(() => {
-          const success = false; // or random
-          setSubmissionStatus({
-            type: success ? "success" : "error",
-            message: success ? "Submission successful!" : "Submission failed. Please try again.",
-          });
-        }, 0);
-      };
       
 
  
     return (
         <div className="flex flex-col items-center">
-        
-        
-            <Card className="w-full max-w-md shadow-lg">
-
-                <CardHeader>
-                    <CardTitle>Placeholder: For Components... <br /> <br /> "Get started now: upload a resume:"</CardTitle>
-                </CardHeader>
-
-                <CardContent>
-                    <CardDescription>
-                        Add all 4 components here: pdf getter, docx getter, user text input, txt file getter:
-                        Arrange in box pattern, 
-                         {/* <img src="/home-page-example.jpg" alt="Fetched Image" /> */}
-                        <br />
-                        <br />
-                        [pdf]  [docx]
-                        <br />
-                        [txt]  [free text]
-
-                    </CardDescription>
-
-                    <button
-            onClick={handleSubmit}
-            className="mt-4 px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700 transition"
-          >
-            Upload
-          </button>
-
-          {submissionStatus && (
-            <div className="mt-4">
-              <SubmissionFeedback
-                type={submissionStatus.type}
-                message={submissionStatus.message}
-              />
-            </div>
-          )}
-                </CardContent>
-
-
-                <CardFooter>
-                    <CardDescription></CardDescription>
-                </CardFooter>
-
-            </Card>
 
 
 {/* ---------------------  Document Input components Here: ------------------------------------------------- */}
