+++ src/app/resume/page.tsx
@@ -5,12 +5,10 @@ import { useAuth } from "@/context/authContext";
 import { useRouter } from "next/navigation";
 import { useState, useEffect } from "react";
 import { Card, CardHeader, CardTitle, CardContent, CardDescription, CardFooter } from "@/components/ui/card";
-
 import { UserNameAddUpdate } from '@/components/userNameAddUpdate';
-
-
-
 import Spinner, { spinnerStyles } from '../../components/ui/Spinner';
+import SubmissionFeedback from "@/components/SubmissionFeedback";
+
 
 
 
@@ -18,6 +16,8 @@ export default function HomePage() {
 
     const { user, loading } = useAuth();
     const router = useRouter();
+    const [submissionStatus, setSubmissionStatus] = useState<null | { type: "success" | "error"; message: string }>(null);
+
 
 
     // Debugging:
@@ -35,11 +35,57 @@ export default function HomePage() {
         return <p>Loading...</p>; // Show a loading state while checking auth
     }
 
-    
+    const handleSubmit = () => {
+        setSubmissionStatus(null); // clear current state first
+      
+        setTimeout(() => {
+          const success = false; // or random
+          setSubmissionStatus({
+            type: success ? "success" : "error",
+            message: success ? "Submission successful!" : "Submission failed. Please try again.",
+          });
+        }, 0);
+    };
 
  
     return (
+        
         <div className="flex flex-col items-center">
+            
+            <Card className="w-full max-w-md shadow-lg">
+            <CardHeader>
+                <CardTitle>
+                Get started now
+                </CardTitle>
+            </CardHeader>
+
+            <CardContent>
+                <CardDescription>
+                Upload a DOCX file containing your resume.
+                </CardDescription>
+
+                <button
+                onClick={handleSubmit}
+                className="mt-4 px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700 transition"
+                >
+                Upload
+                </button>
+
+                {submissionStatus && (
+                <div className="mt-4">
+                    <SubmissionFeedback
+                    type={submissionStatus.type}
+                    message={submissionStatus.message}
+                    />
+                </div>
+                )}
+            </CardContent>
+
+            <CardFooter>
+                <CardDescription></CardDescription>
+            </CardFooter>
+            </Card>
+
 
 
 
@@ -65,6 +111,7 @@ export default function HomePage() {
                 </CardFooter>
 
             </Card>
+
